pyproject.toml
[project]
name = "systematic-search-generator"
version = "0.1.0"
description = "Automated literature retrieval tool that compiles high-quality study lists for systematic reviews"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.86",
    "openpyxl>=3.1.0",
    "openai>=1.0.0",
    "python-docx>=1.0.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
]
